ShiftInMemoryRepository.cs, Shift.cs, IShiftRepository.cs completely done

+++ Entities/Shift.cs
@@ -2,7 +2,7 @@
 
 public class Shift
 {
-    public Shift(int id, DateTime startDateTime, DateTime endDateTime, string typeOfShift, string shiftStatus, string description, string location)
+    public Shift(long id, DateTime startDateTime, DateTime endDateTime, string typeOfShift, string shiftStatus, string description, string location)
     {
         Id = id;
         StartDateTime = startDateTime;
@@ -14,7 +14,7 @@ public class Shift
         
     }
 
-    public int Id { get; set; }
+    public long Id { get; set; }
     public DateTime StartDateTime { get; set; }
     public DateTime EndDateTime { get; set; }
     public string TypeOfShift { get; set; }
